requirements.txt
PySide6>=6.6.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
bcrypt>=4.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
